+++ src/dispatch/model.py
@@ -2,13 +2,11 @@
 from __future__ import annotations
 
 import inspect
-import json
 import logging
 from collections.abc import Callable
 from datetime import datetime
-from io import BytesIO
 from pathlib import Path
-from zipfile import ZIP_DEFLATED, ZipFile
+from zipfile import ZIP_DEFLATED
 
 import numpy as np
 import pandas as pd
@@ -24,8 +22,8 @@ __all__ = ["DispatchModel"]
 
 from dispatch import __version__
 from dispatch.engine import dispatch_engine, dispatch_engine_compiled
-from dispatch.helpers import _null, _str_cols, _to_frame, apply_op_ret_date
-from dispatch.metadata import NET_LOAD_SCHEMA, Validator
+from dispatch.helpers import DataZip, apply_op_ret_date
+from dispatch.metadata import LOAD_PROFILE_SCHEMA, Validator
 
 LOGGER = logging.getLogger(__name__)
 
@@ -84,12 +82,14 @@ class DispatchModel:
     """
 
     __slots__ = (
+        "load_profile",
         "net_load_profile",
         "dispatchable_specs",
         "dispatchable_cost",
         "dispatchable_profiles",
         "storage_specs",
-        "re_profiles",
+        "re_profiles_ac",
+        "re_excess",
         "re_plant_specs",
         "dt_idx",
         "yrs_idx",
@@ -100,21 +100,24 @@ class DispatchModel:
         "_metadata",
         "_cached",
     )
-    _parquet_out = {
-        "re_profiles": _null,
-        "storage_dispatch": _null,
-        "system_data": _null,
-        "storage_specs": _null,
-        "dispatchable_specs": _null,
-        "dispatchable_cost": _null,
-        "dispatchable_profiles": _str_cols,
-        "redispatch": _str_cols,
-        "net_load_profile": _to_frame,
-    }
+    _parquet_out = (
+        "re_profiles_ac",
+        "re_excess",
+        "re_plant_specs",
+        "storage_dispatch",
+        "system_data",
+        "storage_specs",
+        "dispatchable_specs",
+        "dispatchable_cost",
+        "dispatchable_profiles",
+        "redispatch",
+        "load_profile",
+        "net_load_profile",
+    )
 
     def __init__(
         self,
-        net_load_profile: pd.Series[float],
+        load_profile: pd.Series[float],
         dispatchable_specs: pd.DataFrame,
         dispatchable_profiles: pd.DataFrame,
         dispatchable_cost: pd.DataFrame,
@@ -127,8 +130,10 @@ class DispatchModel:
         """Initialize DispatchModel.
 
         Args:
-            net_load_profile: net load, as in net of RE generation, negative net
-                load means excess renewable generation
+            load_profile: load profile that resources will be dispatched against. If
+                ``re_profiles`` and ``re_plant_specs`` are not provided, this should be
+                a net load profile. If they are provided, this *must* be a gross
+                profile, or at least, gross of those RE resources.
             dispatchable_specs: rows are dispatchable generators, columns must include:
 
                 -   capacity_mw: generator nameplate/operating capacity
@@ -136,7 +141,7 @@ class DispatchModel:
                 -   operating_date: the date the plant entered or will enter service
                 -   retirement_date: the date the plant will retire
 
-            dispatchable_profiles: set the maximum output of each generator in each hour
+            dispatchable_profiles: set the maximum output of each generator in each hour.
             dispatchable_cost: cost metrics for each dispatchable generator in each year
                 must be tidy with :class:`pandas.MultiIndex` of
                 ``['plant_id_eia', 'generator_id', 'datetime']``, columns must
@@ -147,15 +152,29 @@ class DispatchModel:
                 -   fom_per_kw: fixed O&M (USD/kW)
                 -   start_per_kw: generator startup cost (USD/kW)
 
-            storage_specs: rows are types of storage, columns must include:
+            storage_specs: rows are storage facilities, for RE+Storage facilities,
+                the ``plant_id_eia`` for the storage component must match the
+                ``plant_id_eia`` for the RE component in ``re_profiles`` and
+                ``re_plant_specs``. Columns must include:
 
                 -   capacity_mw: max charge/discharge capacity in MW
                 -   duration_hrs: storage duration in hours
                 -   roundtrip_eff: roundtrip efficiency
                 -   operating_date: datetime unit starts operating
 
-            re_profiles: ??
-            re_plant_specs: ??
+            re_profiles: normalized renewable profiles, these should be DC profiles,
+                especially when they are part of RE+Storage resources, if they are
+                AC profiles, make sure the ilr in ``re_plant_specs`` is 1.0.
+            re_plant_specs: rows are renewable facilities, for RE+Storage facilities,
+                the ``plant_id_eia`` for the RE component must match the
+                ``plant_id_eia`` for the storage component in ``storage_specs``.
+                Columns must include:
+
+                -   capacity_mw: AC capacity of the generator
+                -   ilr: inverter loading ratio, if ilr != 1, the corresponding profile
+                    must be a DC profile.
+                -   operating_date: datetime unit starts operating
+
             jit: if ``True``, use numba to compile the dispatch engine, ``False`` is
                 mostly for debugging
             name: a name, only used in the ``repr``
@@ -169,9 +188,9 @@ class DispatchModel:
             "jit": jit,
         }
 
-        self.net_load_profile: pd.Series = NET_LOAD_SCHEMA.validate(net_load_profile)
+        self.load_profile: pd.Series = LOAD_PROFILE_SCHEMA.validate(load_profile)
 
-        self.dt_idx = self.net_load_profile.index
+        self.dt_idx = self.load_profile.index
         self.yrs_idx = self.dt_idx.to_series().groupby([pd.Grouper(freq="YS")]).first()
 
         validator = Validator(self, gen_set=dispatchable_specs.index)
@@ -187,16 +206,21 @@ class DispatchModel:
             self.dispatchable_specs.operating_date,
             self.dispatchable_specs.retirement_date,
         )
-        self.re_plant_specs, self.re_profiles = validator.renewables(
+        self.re_plant_specs, re_profiles = validator.renewables(
             re_plant_specs, re_profiles
         )
+        (
+            self.net_load_profile,
+            self.re_profiles_ac,
+            self.re_excess,
+        ) = self.re_and_net_load(re_profiles)
 
         # create vars with correct column names that will be replaced after dispatch
         self.redispatch = MTDF.reindex(columns=self.dispatchable_specs.index)
         self.storage_dispatch = MTDF.reindex(
             columns=[
                 col
-                for i in range(self.storage_specs.shape[0])
+                for i in self.storage_specs.index.get_level_values("plant_id_eia")
                 for col in (f"charge_{i}", f"discharge_{i}", f"soc_{i}")
             ]
         )
@@ -205,6 +229,36 @@ class DispatchModel:
         )
         self.starts = MTDF.reindex(columns=self.dispatchable_specs.index)
         self._cached = {}
+
+    def re_and_net_load(self, re_profiles):
+        """Create net_load_profile based on what RE data was provided."""
+        if self.re_plant_specs is None or re_profiles is None:
+            return (
+                self.load_profile,
+                None,
+                None,
+            )
+        # ILR adjusted normalized profiles
+        temp = re_profiles * self.re_plant_specs.ilr.to_numpy()
+        ac_out = np.minimum(temp, 1) * self.re_plant_specs.capacity_mw.to_numpy()
+        excess = temp * self.re_plant_specs.capacity_mw.to_numpy() - ac_out
+        return self.load_profile - ac_out.sum(axis=1), ac_out, excess
+
+    def dc_charge(self):
+        """Align excess_re to match the storage facilities it could charge."""
+        dc_charge = pd.DataFrame(
+            np.nan, index=self.load_profile.index, columns=self.storage_specs.index
+        )
+        if self.re_excess is None:
+            return dc_charge.fillna(0.0)
+        dc_charge = dc_charge.droplevel("generator_id", axis=1)
+        return (
+            dc_charge.combine_first(self.re_excess.droplevel("generator_id", axis=1))[
+                dc_charge.columns
+            ]
+            .set_axis(self.storage_specs.index, axis=1)
+            .fillna(0.0)
+        )
 
     def add_total_costs(self, df: pd.DataFrame) -> pd.DataFrame:
         """Add columns for total FOM and total startup from respective unit costs."""
@@ -236,19 +290,10 @@ class DispatchModel:
                 )
             del meta["__qualname__"]
 
-        data_dict = {}
-        with ZipFile(path.with_suffix(".zip"), "r") as z:
-            metadata = json.loads(z.read("metadata.json"))
+        with DataZip(path, "r") as z:
+            metadata = z.read("metadata")
             _type_check(metadata)
-            plant_index = pd.MultiIndex.from_tuples(
-                metadata.pop("plant_index"), names=["plant_id_eia", "generator_id"]
-            )
-            for df_name in cls._parquet_out:
-                if (x := df_name + ".parquet") in z.namelist():
-                    df_in = pd.read_parquet(BytesIO(z.read(x))).squeeze()
-                    if df_name in ("dispatchable_profiles", "redispatch"):
-                        df_in.columns = plant_index
-                    data_dict[df_name] = df_in
+            data_dict = dict(z.read_dfs())
 
         sig = inspect.signature(cls).parameters
         self = cls(
@@ -272,9 +317,15 @@ class DispatchModel:
     ) -> DispatchModel:
         """Create :class:`.DispatchModel` with data from patio.BAScenario."""
         if "operating_date" not in storage_specs:
-            storage_specs = storage_specs.assign(operating_date=net_load.index.min())
+            storage_specs = storage_specs.assign(
+                operating_date=net_load.index.min(),
+                plant_id_eia=lambda x: x.index.to_series() * -1,
+                generator_id=lambda x: (
+                    x.groupby(["plant_id_eia"]).transform("cumcount") + 1
+                ).astype(str),
+            ).set_index(["plant_id_eia", "generator_id"])
         return cls(
-            net_load_profile=net_load,
+            load_profile=net_load,
             dispatchable_specs=plant_data,
             dispatchable_cost=cost_data,
             dispatchable_profiles=dispatchable_profiles,
@@ -321,17 +372,16 @@ class DispatchModel:
 
         # make a boolean array for whether a particular hour comes between
         # a generator's `operating_date` and `retirement_date` or not
-        dispatchable_profiles = apply_op_ret_date(
-            pd.DataFrame(
-                1, index=net_load_profile.index, columns=dispatchable_specs.index
+        dispatchable_profiles = pd.DataFrame(
+            np.expand_dims(dispatchable_specs.capacity_mw.to_numpy(), axis=0).repeat(
+                len(net_load_profile.index), axis=0
             ),
-            dispatchable_specs.operating_date,
-            dispatchable_specs.retirement_date,
-            dispatchable_specs.capacity_mw,
+            index=net_load_profile.index,
+            columns=dispatchable_specs.index,
         )
 
         return cls(
-            net_load_profile=net_load_profile,
+            load_profile=net_load_profile,
             dispatchable_specs=dispatchable_specs,
             dispatchable_cost=dispatchable_cost,
             dispatchable_profiles=dispatchable_profiles,
@@ -420,7 +470,7 @@ class DispatchModel:
             index=self.dt_idx,
             columns=[
                 col
-                for i in range(storage.shape[2])
+                for i in self.storage_specs.index.get_level_values("plant_id_eia")
                 for col in (f"charge_{i}", f"discharge_{i}", f"soc_{i}")
             ],
         )
@@ -445,16 +495,16 @@ class DispatchModel:
         profs = profiles.to_numpy()
         fuel_cost = profs * self.dispatchable_cost.fuel_per_mwh.unstack(
             level=("plant_id_eia", "generator_id")
-        ).reindex(index=self.net_load_profile.index, method="ffill")
+        ).reindex(index=self.load_profile.index, method="ffill")
         vom_cost = profs * self.dispatchable_cost.vom_per_mwh.unstack(
             level=("plant_id_eia", "generator_id")
-        ).reindex(index=self.net_load_profile.index, method="ffill")
+        ).reindex(index=self.load_profile.index, method="ffill")
         start_cost = np.where(
             (profs == 0) & (np.roll(profs, -1, axis=0) > 0), 1, 0
         ) * self.dispatchable_cost.startup_cost.unstack(
             level=("plant_id_eia", "generator_id")
         ).reindex(
-            index=self.net_load_profile.index, method="ffill"
+            index=self.load_profile.index, method="ffill"
         )
         fom = (
             apply_op_ret_date(
@@ -468,9 +518,9 @@ class DispatchModel:
                     lambda x: x.replace(day=1, month=1)
                 ),
             )
-            .reindex(index=self.net_load_profile.index, method="ffill")
+            .reindex(index=self.load_profile.index, method="ffill")
             .divide(
-                self.net_load_profile.groupby(pd.Grouper(freq="YS")).transform("count"),
+                self.load_profile.groupby(pd.Grouper(freq="YS")).transform("count"),
                 axis=0,
             )
         )
@@ -567,7 +617,7 @@ class DispatchModel:
     ) -> pd.Series[int]:
         """Number of hours during which deficit was in various duration bins."""
         if comparison is None:
-            durs = self.system_data.deficit / self.net_load_profile
+            durs = self.system_data.deficit / self.load_profile
         else:
             durs = self.system_data.deficit / comparison
         bins = map(
@@ -590,9 +640,7 @@ class DispatchModel:
         positive deficit hours.
         """
         if comparison is None:
-            comparison = self.net_load_profile.groupby(
-                [pd.Grouper(freq="YS")]
-            ).transform(
+            comparison = self.load_profile.groupby([pd.Grouper(freq="YS")]).transform(
                 "max"
             )  # type: ignore
         td_1h = np.timedelta64(1, "h")
@@ -603,9 +651,43 @@ class DispatchModel:
                     self.system_data.deficit / comparison > cutoff
                 ].index
                 for hr in (dhr - 2 * td_1h, dhr - td_1h, dhr)
-                if hr in self.net_load_profile.index
+                if hr in self.load_profile.index
             }
         )
+
+    def hourly_data_check(self, cutoff: float = 0.01):
+        """Aggregate data for :meth:`.DispatchModel.hrs_to_checl`."""
+        max_disp = apply_op_ret_date(
+            pd.DataFrame(
+                1.0,
+                index=self.load_profile.index,
+                columns=self.dispatchable_profiles.columns,
+            ),
+            self.dispatchable_specs.operating_date,
+            self.dispatchable_specs.retirement_date,
+            self.dispatchable_specs.capacity_mw,
+        )
+        out = pd.concat(
+            {
+                "gross_load": self.load_profile,
+                "net_load": self.net_load_profile,
+                "deficit": self.system_data.deficit,
+                "max_dispatch": max_disp.sum(axis=1),
+                "redispatch": self.redispatch.sum(axis=1),
+                "historical_dispatch": self.dispatchable_profiles.sum(axis=1),
+                "net_storage": (
+                    self.storage_dispatch.filter(regex="^discharge").sum(axis=1)
+                    - self.storage_dispatch.filter(regex="^charge").sum(axis=1)
+                ),
+                "state_of_charge": self.storage_dispatch.filter(regex="^soc").sum(
+                    axis=1
+                ),
+                "re": self.re_profiles_ac.sum(axis=1),
+                "re_excess": self.re_excess.sum(axis=1),
+            },
+            axis=1,
+        ).loc[self.hrs_to_check(cutoff=cutoff), :]
+        return out
 
     def storage_capacity(self) -> pd.DataFrame:
         """Number of hours when storage charge or discharge was in various bins."""
@@ -648,17 +730,15 @@ class DispatchModel:
                 self.system_data.groupby(pd.Grouper(freq=freq))
                 .sum()
                 .rename(columns={c: f"{c}_mwh" for c in self.system_data}),
-                # max deficit pct of net load
+                # max deficit pct of load
                 self.system_data[["deficit"]]
                 .groupby(pd.Grouper(freq=freq))
                 .max()
                 .rename(columns={"deficit": "deficit_max_pct_net_load"})
-                / self.net_load_profile.max(),
+                / self.load_profile.max(),
                 # count of deficit greater than 2%
                 pd.Series(
-                    self.system_data[
-                        self.system_data / self.net_load_profile.max() > 0.02
-                    ]
+                    self.system_data[self.system_data / self.load_profile.max() > 0.02]
                     .groupby(pd.Grouper(freq=freq))
                     .deficit.count(),
                     name="deficit_gt_2pct_count",
@@ -703,12 +783,12 @@ class DispatchModel:
         **kwargs,
     ) -> pd.DataFrame:
         """Create granular summary of renewable plant metrics."""
-        if self.re_profiles is None or self.re_plant_specs is None:
+        if self.re_profiles_ac is None or self.re_plant_specs is None:
             raise AssertionError(
                 "at least one of `re_profiles` and `re_plant_specs` is `None`"
             )
         out = (
-            self.re_profiles.groupby([pd.Grouper(freq=freq)])
+            self.re_profiles_ac.groupby([pd.Grouper(freq=freq)])
             .sum()
             .stack(["plant_id_eia", "generator_id"])
             .to_frame(name="redispatch_mwh")
@@ -744,19 +824,28 @@ class DispatchModel:
             .rename(columns={0: "redispatch_mwh"})
         )
 
-        out[["kind", "index"]] = out.level_1.str.split("_", expand=True)
+        out[["kind", "plant_id_eia"]] = out.level_1.str.split("_", expand=True)
         out = (
             out.query("kind != 'soc'")
+            .astype({"plant_id_eia": int})
             .assign(
                 redispatch_mwh=lambda x: x.redispatch_mwh.mask(
                     x.kind == "charge", x.redispatch_mwh * -1
-                )
+                ),
+                generator_id=lambda x: x.plant_id_eia.replace(
+                    self.storage_specs.reset_index(
+                        "generator_id"
+                    ).generator_id.to_dict()
+                ),
             )
-            .groupby(["index", "datetime"])
+            .groupby(["plant_id_eia", "generator_id", "datetime"])
             .redispatch_mwh.sum()
             .reset_index()
-            .astype({"index": int})
-            .merge(self.storage_specs.reset_index(), on="index", validate="m:1")
+            .merge(
+                self.storage_specs.reset_index(),
+                on=["plant_id_eia", "generator_id"],
+                validate="m:1",
+            )
             .assign(
                 capacity_mw=lambda x: x.capacity_mw.where(
                     x.operating_date <= x.datetime, 0
@@ -764,14 +853,7 @@ class DispatchModel:
             )
         )
         if by is None:
-            if (
-                None in self.storage_specs.index.names
-                and len(self.storage_specs.index.names) == 1
-            ):
-                col = ["index"]
-            else:
-                col = self.storage_specs.index.names
-            return out.set_index(col + ["datetime"])
+            return out.set_index(["plant_id_eia", "generator_id", "datetime"])
         return out.groupby([by, "datetime"]).sum()
 
     def full_output(self, freq: str = "YS") -> pd.DataFrame:
@@ -805,13 +887,63 @@ class DispatchModel:
                 + [col for col in cols if col in self.dispatchable_specs]
             ]
         )
+
+        # setup deficit/curtailment as if they were resources for full output, the idea
+        # here is that you could rename them purchase/sales.
+        def_cur = self.grouper(self.system_data, by=None)[["deficit", "curtailment"]]
+        def_cur.columns = pd.MultiIndex.from_tuples(
+            [(0, "deficit"), (0, "curtailment")], names=["plant_id_eia", "generator_id"]
+        )
+        def_cur = (
+            def_cur.stack([0, 1])
+            .reorder_levels([1, 2, 0])
+            .sort_index()
+            .to_frame(name="redispatch_mwh")
+            .assign(
+                technology_description=lambda x: x.index.get_level_values(
+                    "generator_id"
+                )
+            )
+        )
+
         return pd.concat(
             [
                 dispatchable,
                 self.re_summary(by=None, freq=freq),
-                self.storage_summary(by=None, freq=freq),
+                self.storage_summary(by=None, freq=freq)
+                .reset_index()
+                .set_index(["plant_id_eia", "generator_id", "datetime"]),
+                def_cur,
             ]
         ).sort_index()
+
+    def load_summary(self, **kwargs):
+        """Create summary of load data."""
+        return pd.concat(
+            [
+                self.strict_grouper(
+                    self.net_load_profile.to_frame("net_load"), by=None, freq="YS"
+                ),
+                self.strict_grouper(
+                    self.net_load_profile.to_frame("net_load_peak"),
+                    by=None,
+                    freq="YS",
+                    freq_agg="max",
+                ),
+                self.strict_grouper(
+                    self.load_profile.to_frame("gross_load"),
+                    by=None,
+                    freq="YS",
+                ),
+                self.strict_grouper(
+                    self.load_profile.to_frame("gross_load_peak"),
+                    by=None,
+                    freq="YS",
+                    freq_agg="max",
+                ),
+            ],
+            axis=1,
+        )
 
     def dispatchable_summary(
         self,
@@ -833,7 +965,7 @@ class DispatchModel:
                         apply_op_ret_date(
                             pd.DataFrame(
                                 1,
-                                index=self.net_load_profile.index,
+                                index=self.load_profile.index,
                                 columns=self.dispatchable_specs.index,
                             ),
                             self.dispatchable_specs.operating_date,
@@ -900,42 +1032,25 @@ class DispatchModel:
         clobber=False,
         **kwargs,
     ) -> None:
-        """Save :class:`.DispatchModel` to disk.
-
-        A very ugly process at the moment because of our goal not to use pickle
-        and to try to keep the file small-ish. Also need to manage the fact that
-        the parquet requirement for string column names causes some problems.
-        """
-        if not isinstance(path, Path):
-            path = Path(path)
-        path = path.with_suffix(".zip")
-        if path.exists() and not clobber:
+        """Save :class:`.DispatchModel` to disk."""
+        if Path(path).with_suffix(".zip").exists() and not clobber:
             raise FileExistsError(f"{path} exists, to overwrite set `clobber=True`")
 
-        metadata = {
-            **self._metadata,
-            "__qualname__": self.__class__.__qualname__,
-            "plant_index": list(self.dispatchable_specs.index),
-        }
-
-        with ZipFile(path, "w", compression=compression) as z:
-            for df_name, func in self._parquet_out.items():
-                try:
-                    df_out = getattr(self, df_name)
-                    if df_out is not None and not df_out.empty:
-                        z.writestr(
-                            f"{df_name}.parquet", func(df_out, df_name).to_parquet()
-                        )
-                except Exception as exc:
-                    raise RuntimeError(f"{df_name} {exc!r}") from exc
+        with DataZip(path, "w", compression=compression) as z:
+            for df_name in self._parquet_out:
+                z.writed(df_name, getattr(self, df_name))
             if include_output and not self.redispatch.empty:
-                for df_name in ("system_level_summary", "dispatchable_summary"):
-                    z.writestr(
-                        f"{df_name}.parquet",
-                        getattr(self, df_name)(**kwargs).to_parquet(),
+                for df_name in ("full_output", "load_summary"):
+                    z.writed(
+                        df_name,
+                        getattr(self, df_name)(**kwargs),
                     )
-            z.writestr(
-                "metadata.json", json.dumps(metadata, ensure_ascii=False, indent=4)
+            z.writed(
+                "metadata",
+                {
+                    **self._metadata,
+                    "__qualname__": self.__class__.__qualname__,
+                },
             )
 
     def _plot_prep(self):
@@ -947,7 +1062,7 @@ class DispatchModel:
             try:
                 re = self.re_summary(freq="H").redispatch_mwh.unstack(level=0)
             except AssertionError:
-                re = MTDF.reindex(index=self.net_load_profile.index)
+                re = MTDF.reindex(index=self.load_profile.index)
 
             def _grp(df):
                 return df.rename(columns=PLOT_MAP).groupby(level=0, axis=1).sum()
@@ -973,7 +1088,7 @@ class DispatchModel:
 
     def plot_period(self, begin, end):
         """Plot hourly dispatch."""
-        net_load = self.net_load_profile.loc[begin:end]
+        net_load = self.load_profile.loc[begin:end]
         out = px.bar(
             self._plot_prep().loc[net_load.index, :].reset_index(),
             x="datetime",
@@ -988,9 +1103,9 @@ class DispatchModel:
             line_color=COLOR_MAP["Net Load"],
             line_dash="dot",
         )
-        if self.re_profiles is None or self.re_plant_specs is None:
+        if self.re_profiles_ac is None or self.re_plant_specs is None:
             return out
-        gross = self.re_profiles.loc[net_load.index, :].sum(axis=1) + net_load
+        gross = self.re_profiles_ac.loc[net_load.index, :].sum(axis=1) + net_load
         return out.add_scatter(
             x=gross.index,
             y=gross,